+++ src/errors.rs
@@ -20,8 +20,9 @@ pub enum APIError {
     #[error("License error: {0}")]
     License(String),
 
-    // #[error("Not accessible")]
-    // NotAccessible(),
+    #[error("Not accessible")]
+    NotAccessible(),
+
     #[error("Unauthorized registration access")]
     Unauthorized(),
 
@@ -40,14 +41,6 @@ impl IntoResponse for APIError {
         error!("API Error occurred: {}", self);
 
         let (status_code, error_response) = match &self {
-            // APIError::Unexpected(_) => (
-            //     StatusCode::INTERNAL_SERVER_ERROR,
-            //     ApiError {
-            //         status: "failed".to_string(),
-            //         reason: "Internal Server Error".to_string(),
-            //         details: "Please contact our support at https://blockfrost.io".to_string(),
-            //     },
-            // ),
             APIError::Validaion(_) => (
                 StatusCode::BAD_REQUEST,
                 ApiError {
@@ -64,14 +57,14 @@ impl IntoResponse for APIError {
                     details: format!("Address: {} does not contain the license.", address),
                 },
             ),
-            // APIError::NotAccessible() => (
-            //     StatusCode::FORBIDDEN,
-            //     ApiError {
-            //         status: "failed".to_string(),
-            //         reason: "not_accessible".to_string(),
-            //         details: "The Blockfrost instance is not publicly accessible.".to_string(),
-            //     },
-            // ),
+            APIError::NotAccessible() => (
+                StatusCode::FORBIDDEN,
+                ApiError {
+                    status: "failed".to_string(),
+                    reason: "not_accessible".to_string(),
+                    details: "The Blockfrost instance is not publicly accessible.".to_string(),
+                },
+            ),
             APIError::Unauthorized() => (
                 StatusCode::FORBIDDEN,
                 ApiError {